requirements.txt
opencv-python
tqdm
numpy
pandas
torch
ultralytics
matplotlib
scipy
Pillow
numpy
tensorboard
pyyaml
yacs